Constructing new module dependencies
