+++ js/modules/templates/game-2.js
@@ -42,14 +42,11 @@ const moduleGame2 = `<header class="header">
       </ul>
     </div>
   </div>`;
-const templateGame2 = (template) =>{
+export const templateGame2 = (template) =>{
   const container = document.createElement(`template`);
   container.id = `game-2`;
   container.innerHTML = moduleGame2;
   document.querySelector(`#game-1`).after(container);
 };
-templateGame2(moduleGame2);
-
-export default templateGame2;
