+++ js/modules/templates/stats.js
@@ -106,13 +106,9 @@ const moduleStats = `<header class="header">
       </tr>
     </table>
   </div>`;
-const templateStats = (template) =>{
+export const templateStats = (template) =>{
   const container = document.createElement(`template`);
   container.id = `stats`;
   container.innerHTML = moduleStats;
   document.querySelector(`#game-3`).after(container);
 };
-templateStats(moduleStats);
-
-export default templateStats;
-
